services/agent/requirements.txt
nanobot-ai
aiohttp
loguru
